requirements.txt
apify_client
crawl4ai
openai
python-dotenv
aiohttp
urllib3
fastapi
uvicorn
httpx
python-dotenv
pydantic
supabase
asyncpg
bcrypt
python-jose[cryptography]
passlib
python-jose
pydantic[email]
python-multipart
apscheduler
serpapi
httpx
requests
beautifulsoup4
lxml
pathlib2
pandas
loguru
praw
requests 
requests_oauthlib
langchain
langchain-community
google-generativeai
langdetect
selenium
beautifulsoup4
lxml
webdriver-manager
reportlab
